+++ src/utils/shared_memory.cpp
@@ -15,7 +15,7 @@ namespace utils {
 SharedMemory::SharedMemory(std::string_view base_path)
     : base_path_(base_path),
       mmap_delay_stat_(
-          stat::StatisticsCollector<uint32_t>::StandardReportCallback("mmap_delay")) {
+          stat::StatisticsCollector<int32_t>::StandardReportCallback("mmap_delay")) {
     CHECK(fs_utils::IsDirectory(base_path_));
 }
 
@@ -30,7 +30,7 @@ SharedMemory::~SharedMemory() {
 
 SharedMemory::Region* SharedMemory::Create(std::string_view path, size_t size) {
     std::string full_path = GetFullPath(path);
-    uint64_t start_timestamp = GetMonotonicMicroTimestamp();
+    int64_t start_timestamp = GetMonotonicMicroTimestamp();
     int fd = open(full_path.c_str(), O_CREAT|O_EXCL|O_RDWR, 0644);
     PCHECK(fd != -1) << "open failed";
     PCHECK(ftruncate(fd, size) == 0) << "ftruncate failed";
@@ -40,7 +40,8 @@ SharedMemory::Region* SharedMemory::Create(std::string_view path, size_t size) {
         PCHECK(ptr != MAP_FAILED) << "mmap failed";
         PCHECK(close(fd) == 0) << "close failed";
         memset(ptr, 0, size);
-        mmap_delay_stat_.AddSample(GetMonotonicMicroTimestamp() - start_timestamp);
+        mmap_delay_stat_.AddSample(gsl::narrow_cast<int32_t>(
+            GetMonotonicMicroTimestamp() - start_timestamp));
     }
     Region* region = new Region(this, path, reinterpret_cast<char*>(ptr), size);
     AddRegion(region);
@@ -49,7 +50,7 @@ SharedMemory::Region* SharedMemory::Create(std::string_view path, size_t size) {
 
 SharedMemory::Region* SharedMemory::OpenReadOnly(std::string_view path) {
     std::string full_path = GetFullPath(path);
-    uint64_t start_timestamp = GetMonotonicMicroTimestamp();
+    int64_t start_timestamp = GetMonotonicMicroTimestamp();
     int fd = open(full_path.c_str(), O_RDONLY);
     PCHECK(fd != -1) << "open failed";
     struct stat statbuf;
@@ -60,7 +61,8 @@ SharedMemory::Region* SharedMemory::OpenReadOnly(std::string_view path) {
         ptr = mmap(0, size, PROT_READ, MAP_SHARED, fd, 0);
         PCHECK(ptr != MAP_FAILED) << "mmap failed";
         PCHECK(close(fd) == 0) << "close failed";
-        mmap_delay_stat_.AddSample(GetMonotonicMicroTimestamp() - start_timestamp);
+        mmap_delay_stat_.AddSample(gsl::narrow_cast<int32_t>(
+            GetMonotonicMicroTimestamp() - start_timestamp));
     }
     Region* region = new Region(this, path, reinterpret_cast<char*>(ptr), size);
     AddRegion(region);
